# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Gastronómico",
    "version": "1.0.0",
    "description": "Documentación de la API Multi-Tenant para Restaurantes",
    "contact": {
      "name": "Soporte API",
      "email": "api@gastronomico.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/v1",
      "description": "Servidor local"
    },
    {
      "url": "https://backend-gastronomico.onrender.com/v1",
      "description": "Servidor desplegado"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtenido del endpoint /auth/login"
      }
    }
  },
...
  "paths": {
...
    "/contact": {
      "post": {
        "summary": "Crear un nuevo contacto",
        "description": "Crea un contacto que vincula un cliente con una instancia (canal de comunicación).",
        "tags": [
          "Contact"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "customerId",
                  "instanceId",
                  "storeId",
                  "identifier",
                  "type"
                ],
                "properties": {
                  "customerId": {
                    "type": "integer",
                    "description": "ID del cliente"
                  },
                  "instanceId": {
                    "type": "integer",
                    "description": "ID de la instancia (canal - WhatsApp, email, etc.)"
                  },
                  "storeId": {
                    "type": "integer",
                    "description": "ID de la tienda"
                  },
                  "identifier": {
                    "type": "string",
                    "description": "Identificador único (teléfono, email, usuario de WhatsApp, etc.)"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "phone",
                      "email",
                      "whatsapp",
                      "telegram",
                      "other"
                    ],
                    "description": "Tipo de contacto"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contacto creado exitosamente"
          },
          "400": {
            "description": "Validación fallida o contacto duplicado"
          },
          "404": {
            "description": "Cliente o instancia no encontrado"
          }
        }
      }
    },
    "/contact/customer/{customerId}": {
      "get": {
        "summary": "Obtener todos los contactos de un cliente",
        "description": "Retorna todos los canales de comunicación (contactos) asociados a un cliente.",
        "tags": [
          "Contact"
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "description": "ID del cliente",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contactos encontrados"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      }
    },
    "/contact/instance/{instanceId}": {
      "get": {
        "summary": "Obtener todos los contactos de una instancia",
        "description": "Retorna todos los clientes que están en comunicación a través de una instancia específica.",
        "tags": [
          "Contact"
        ],
        "parameters": [
          {
            "name": "instanceId",
            "in": "path",
            "required": true,
            "description": "ID de la instancia",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contactos encontrados"
          },
          "404": {
            "description": "Instancia no encontrada"
          }
        }
      }
    },
    "/contact/{id}": {
      "get": {
        "summary": "Obtener un contacto específico",
        "description": "Retorna los detalles de un contacto.",
        "tags": [
          "Contact"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del contacto",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contacto encontrado"
          },
          "404": {
            "description": "Contacto no encontrado"
          }
        }
      },
      "patch": {
        "summary": "Actualizar un contacto",
        "description": "Actualiza los datos de un contacto existente.",
        "tags": [
          "Contact"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del contacto",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "phone",
                      "email",
                      "whatsapp",
                      "telegram",
                      "other"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contacto actualizado"
          },
          "404": {
            "description": "Contacto no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un contacto",
        "description": "Elimina el contacto entre un cliente y una instancia.",
        "tags": [
          "Contact"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del contacto",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contacto eliminado"
          },
          "404": {
            "description": "Contacto no encontrado"
          }
        }
      }
    },
    "/contact/{id}/status": {
      "patch": {
        "summary": "Cambiar estado de un contacto",
        "description": "Activa o desactiva un contacto.",
        "tags": [
          "Contact"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del contacto",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "statusId"
                ],
                "properties": {
                  "statusId": {
                    "type": "integer",
                    "description": "ID del estado"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado"
          },
          "404": {
            "description": "Contacto no encontrado"
          }
        }
      }
    },
...
    "/customer/{id}": {
      "get": {
        "summary": "Obtener un cliente por ID",
        "tags": [
          "Customer"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      },
      "patch": {
        "summary": "Actualizar un cliente",
        "tags": [
          "Customer"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstname": {
                    "type": "string"
                  },
                  "lastname": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente actualizado"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un cliente",
        "tags": [
          "Customer"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente eliminado"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      }
    },
...
    "/instance/{id}": {
      "get": {
        "summary": "Obtener una instancia por ID",
        "tags": [
          "Instance"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Instancia encontrada"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Instancia no encontrada"
          }
        }
      },
      "patch": {
        "summary": "Actualizar una instancia",
        "tags": [
          "Instance"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "identifier": {
                    "type": "string"
                  },
                  "credentials": {
                    "type": "object"
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Instancia actualizada"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Instancia no encontrada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar una instancia",
        "tags": [
          "Instance"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Instancia eliminada"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Instancia no encontrada"
          }
        }
      }
    },
...
  }
}